Add tests for shadows
