+++ packages/material-ui-system/src/style.test.js
@@ -63,24 +63,6 @@ describe('style', () => {
     // });
   });
 
-  const boxShadow = style({
-    prop: 'boxShadow',
-    themeKey: 'shadows',
-  });
-
-  it('should support array theme value', () => {
-    const output = boxShadow({
-      theme: {
-        shadows: ['none', '0px 1px 3px 0px rgba(0, 0, 0, 0.2)'],
-      },
-      boxShadow: 1,
-    });
-
-    assert.deepEqual(output, {
-      boxShadow: '0px 1px 3px 0px rgba(0, 0, 0, 0.2)',
-    });
-  });
-
   const border = style({
     prop: 'border',
     themeKey: 'borders',
